contact times completed

+++ medguardian/prescriptions/views.py
@@ -27,8 +27,10 @@ from .models import PatientPrescribers
 from .models import Prescription
 from .models import PrescriptionAdminTime
 from .models import AdministrationTime
+from .models import ContactTimes
 from .serializers import PrescriberSerializer
 from .serializers import PrescriptionSerializer
+from src.user_model import Patient
 
 
 class MedGuardianViewMixin(LoginRequiredMixin, UserPassesTestMixin, FormView):
@@ -264,43 +266,93 @@ class PrescriptionUpdateAPIView(LoginRequiredMixin, UserPassesTestMixin, generic
         return self.request.user.id == self.kwargs['pk']
 
     def form_valid(self, form):
-        return None
+        raise Http404()
 
     def get_object(self):
-
         return Prescription.objects.get(id=self.kwargs.get('rx_id'))
 
     def get_serializer(self):
         return None
 
     def patch(self, request, *args, **kwargs):
+        '''
+            In essence a patch on prescription, wherein the administration
+            times are changed (created, modified, deleted).  However, the
+            action is on the association table PrescriptionAdminTimes.
+
+            This also maintains the proper accosicatuons between the contact
+            and administration times through the prescriptions_contacttimes
+            table.  Because multiple prescriptions could have particular
+            administration times, there is an incrementor field, called
+            instances.  When first created this is set to 1.  When decremented
+            to zero, the association is deleted.
+        '''
+
         admin_times = json.loads(self.request.body).get('administration_times', None)
 
-        # if there are new admin times, delete old associations and create new ones
+        # if there are admin times sent, make sure they are different,
+        # then delete old associations and create new ones
         if admin_times:
             # get a list of current administration time ids (if any)
             current_admin_time_ids = self.get_object().administration_times.values_list('id')
             current_admin_time_ids = [time_tup[0] for time_tup in current_admin_time_ids]
 
-            # delete associations with old ids
-            PrescriptionAdminTime.objects.filter(administration_time_id__in=current_admin_time_ids,
-                                                 prescription_id=kwargs.get('rx_id')).delete()
-
             # get ids for new administration times
             admin_time_ids = AdministrationTime.objects.filter(value__in=admin_times).values_list('id')
             admin_time_ids = [time_tup[0] for time_tup in admin_time_ids]
 
-            # create list on new association models objects
-            new_rx_time_associations = [PrescriptionAdminTime(prescription_id=self.kwargs.get('rx_id'),
-                                                              administration_time_id=time_id) for \
-                                                              time_id in admin_time_ids]
+            # only do something if new and old admin times are not the same
+            if sorted(current_admin_time_ids) != sorted(admin_time_ids):
 
-            # create (in bulk) new associations
-            PrescriptionAdminTime.objects.bulk_create(new_rx_time_associations)
+                # delete associations with old ids
+                PrescriptionAdminTime.objects.filter(administration_time_id__in=current_admin_time_ids,
+                                                     prescription_id=kwargs.get('rx_id')).delete()
 
-            ###########NEED TO DELETE OLD ASSOCIATED CONTACT TIMES AND MAKE NEW ASSOCIATIONS
+                # create list on new association models objects
+                new_rx_time_associations = [PrescriptionAdminTime(prescription_id=self.kwargs.get('rx_id'),
+                                                                  administration_time_id=time_id) for \
+                                                                  time_id in admin_time_ids]
+
+                # create (in bulk) new associations
+                PrescriptionAdminTime.objects.bulk_create(new_rx_time_associations)
+
+                # get contact info id for patient
+                self.resolve_contact_times(request.user.id, current_admin_time_ids, admin_time_ids)
 
         return HttpResponse(status=HTTPStatus.ACCEPTED)
 
-    def resolve_contact_times(self, patient_id: int, rx: Prescription, old_time_ids: List[int], new_time_ids: List[int]):
+    def resolve_contact_times(self, patient_id: int, old_time_ids: List[int], new_time_ids: List[int]):
+        '''
+            Remove contact times that are no longer associated with this patient.
+            Make new associations, as appropriate for the new contact times
+        '''
+        contact_id = Patient.objects.get(id=patient_id).contact_information.id
+
+        # get all the old contact times associated with the patient,
+        # but exclude the new times
+        contact_times = ContactTimes.objects.filter(contact_id=contact_id,
+                                                    id__in=old_time_ids).exclude(id__in=new_time_ids)
+
+        # decrement instances column for each removed contact time
+        # if zero, add id to list to delete
+        # ids_to_delete = []
+        for time in contact_times:
+            time.instances -= 1
+            # if time.instances == 0:
+            #     ids_to_delete.append(time.pk)
+
+        # delete contact times with zero instances
+        contact_times.filter(instances=0).delete()
+
+        # update decremented contact times
+        ContactTimes.objects.bulk_update(contact_times, ['instances',])
+
+        # make new contact times if not already existing
+        new_contact_times = [ContactTimes(contact_id=contact_id, administration_time_id=time_id) for \
+                             time_id in new_time_ids if time_id not in old_time_ids]
+
+        # store new association in database
+        ContactTimes.objects.bulk_create(new_contact_times)
+
+
